emmm

+++ src/flame/common/system.cpp
@@ -236,7 +236,7 @@ namespace flame
 	std::unique_ptr<unsigned char[]> read_process_memory(void *process, void *address, int size)
 	{
 		auto m = std::make_unique<unsigned char[]>(size);
-		unsigned long ret_byte;
+		SIZE_T ret_byte;
 		assert(ReadProcessMemory(process, address, m.get(), size, &ret_byte));
 		return m;
 	}
